Set config in config loader
