+++ packages/playwright-test/src/common/configLoader.ts
@@ -22,6 +22,7 @@ import type { ConfigCLIOverrides, SerializedConfig } from './ipc';
 import { requireOrImport } from './transform';
 import type { Config, FullConfigInternal, FullProjectInternal, Project, ReporterDescription } from './types';
 import { errorWithFile, getPackageJsonPath, mergeObjects } from '../util';
+import { setCurrentConfig } from './globals';
 
 export const defaultTimeout = 30000;
 
@@ -47,11 +48,13 @@ export class ConfigLoader {
       throw new Error('Cannot load two config files');
     const config = await requireOrImportDefaultObject(file) as Config;
     await this._processConfigObject(config, path.dirname(file), file);
+    setCurrentConfig(this._fullConfig);
     return this._fullConfig;
   }
 
   async loadEmptyConfig(configDir: string): Promise<Config> {
     await this._processConfigObject({}, configDir);
+    setCurrentConfig(this._fullConfig);
     return {};
   }
