+++ tests/playwright-test/store.spec.ts
@@ -257,22 +257,3 @@ test('store root can be changed with TestConfig.storeDir', async ({ runInlineTes
   const file = path.join(test.info().outputPath(), 'my/store/dir/foo/bar.json');
   expect(JSON.parse(await fs.promises.readFile(file, 'utf-8'))).toEqual({ 'a': 2023 });
 });
-
-test('fulfill with return path of the entry', async ({ runInlineTest, page }) => {
-  const storeDir = path.join(test.info().outputPath(), 'playwright');
-  const file = path.join(storeDir, 'foo/body.json');
-  await fs.promises.mkdir(path.dirname(file), { recursive: true });
-  await fs.promises.writeFile(file, JSON.stringify({ 'a': 2023 }));
-  const result = await runInlineTest({
-    'a.test.ts': `
-      import { test, store, expect } from '@playwright/test';
-      test('should read value from path', async ({ page }) => {
-        await page.route('**/*', route => route.fulfill({ path: store.path('foo/body.json')}))
-        await page.goto('http://example.com');
-        expect(await page.textContent('body')).toBe(JSON.stringify({ 'a': 2023 }))
-      });
-    `,
-  }, { workers: 1 });
-  expect(result.exitCode).toBe(0);
-  expect(result.passed).toBe(1);
-});
